+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add optional `output.localeCodesModule` config file setting which generates a
+  TypeScript module that exports `sourceLocale`, `targetLocales`, and
+  `allLocales` using the locale codes from your config file. Use for keeping
+  your config file and client config in sync.
+
 ### Changed
 
 - **[BREAKING]** Published module paths have changed: